Adding delz fitting parameter to error checks and m-files
This parameter is used when FITDELZ=T to introduce perturbations
to the Grad-Shafranov equation which have Z position dependence.
So the equilibium produced will not be in force balance unless this
quantity vanishes. The residual (ERROR) is pretty good at capturing
this, but it's worth having checks outputs to verify it.

For the new error check (IERCHK>0) the change in axis position (in
meters) is compared against the residual tolerance specified by
ERROR. There's no physics justification for re-using the tolerance,
but it seems like a reasonable target for convergence. A different
tolerance parameter can be added once there is better explanation
of how it should be set. Test results were updated to include the
error check warning when required.

The history of delz with iteration number fit in EFIT is now output
as CDELZ in m-files. (Note it may not have been updated in the last
iteration if EFIT exits the fitting loop before recomputing the
response matrix).

Some additional documentation improvements have also been included.