Fix bound method when lcfs intersects the limiter at search point
This has been the cause of some sporadic EFIT failures.  It seems
to happen because zlim does not handle points near the limiter well.
This fix avoids calling zlim at the first point in these cases, but
doesn't resolve the root issue.  This seems to be at least the third
attempt at a bandaid.

Also doing some more cleanup of the kbound option that isn't being
used.

Serveral output file updates are included that were missed in
previous MR.