Combine separate config scripts for each system
This eliminates a bunch of duplicate code for building with different
compilers on each system. The scripts are now smart enough to detect
which modules are loaded and use the appropriate settings.

This was relatively straight forward for NERSC and ALCF, since the
PrgEnvs there are quite consistent. It was a bit messier for Omega
since those envs are only partially consistent (and there are
additional complications). It is not very feasible for Iris because
nothing is consistent there. That system is being retired anyway.

This also adds instructions for running tests on Polaris.

Some code modernization was performed in the measurements source
file as well.