Generalize the limiter segment selection in gap calculations
Rather than using hardcoded values based on DIII-D, the limits can
be set in the mhdin.dat files and the default option is choose
values based on the limiter dimensions (using scaled thresholds from
DIII-D).

The limits needed to change for the current DIII-D limiter due to
changes to the inboard side. The newly added machine description
file includes this change.

Since the default gap calculations used with ITER and NSTX changed
(which could be set more appropriately in the mhdin.dat files now)
the test results needed to be updated accordingly.