Minor: more clear indication of function definitions
Fortran doesn't require function definitions to be declared any
special way when called in a different subroutine, which makes it
more difficult to read the code. Labeling them as EXTERNAL is only
required when they are used as arguements, but is far easier to
read. So these were added for that purpose.