Add the spline coefficients to the knot opt output files
The first line of the file now communicates the length of the data
arrays in the following lines.

The coefficients for each profile are saved, but evaluated profiles are not,
since they don't actually get constructed until post-processing
(it wouldn't be too bad to hack something in to compute those, but
handling all possible cases requires a lot more code).

This also fixes a bug in the array indexing that was mixing up the
knot positions.