Fix to allow EFIT to run MPI with more cpus than time slices
The extra cpus don't do anything, so an error message is still
printed when this is encountered.

A bug with writing the n-files due to variable name confusion has
been fixed along with the variable descriptions.

Also changing the convergence check messages to begin with "Failure"
instead of "Error" since these are commonly being confused with
execution errors that begin with "ERROR".