Add warnings for NPROC!=4
The test output has only been provided for NPROC=4 so any changes
will cause the tests to fail. It wouldn't be too difficult to
make the tests pass for NPROC>4 (the extra processors will be
ignored by EFIT since only 4 times/k-files are normally provided),
but alternative terminal output files will be needed for NPROC<4
for almost every test, which is more trouble to setup/maintain
than it would be worth. So instead a warning has been added that
changing NPROC will cause tests to fail. The option still may be
useful for developers who are willing to review the test outputs
by hand.

This configuration option has also been removed from the install
instructions since it is not described in detail and should not
be changed general users.