Change default nw_sub and nh_sub values
Rather than defaulting to the chosen resolution nw and nh, these
parameters are now ignored by default. This prevents the resolution
used to create k-files in snap mode from impacting the behavior if
EFIT is run using those k-files at a different resolution (unless
nw_sub and nh_sub are specifically set in the snapfile).

Also fixing the configuration for gnu11 builds on Omega since the
MSE library compiled with gnu8 is compatible.